third_party/WebKit/LayoutTests/fast/multicol/vertical-lr/offset-top-and-left-at-boundaries.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; writing-mode:vertical-lr; columns:3; column-gap:10px; height:320px;">
    <div id="first" style="width:100px;"></div>
    <div id="second" style="width:100px;"></div>
    <div id="third" style="width:100px;"></div>
    <div id="spanner" style="column-span:all; width:50px;"></div>
    <div id="fifth" style="width:100px;"></div>
    <div id="sixth" style="width:100px;"></div>
    <div id="seventh" style="width:100px;"></div>
</div>

<script>
    test(function() {
        assert_equals(document.getElementById('first').offsetLeft, 0);
        assert_equals(document.getElementById('first').offsetTop, 0);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, first element");

    test(function() {
        assert_equals(document.getElementById('second').offsetLeft, 0);
        assert_equals(document.getElementById('second').offsetTop, 110);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, second element");

    test(function() {
        assert_equals(document.getElementById('third').offsetLeft, 0);
        assert_equals(document.getElementById('third').offsetTop, 220);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, third element");

    test(function() {
        assert_equals(document.getElementById('spanner').offsetLeft, 100);
        assert_equals(document.getElementById('spanner').offsetTop, 0);
    }, "offsetLeft and offsetTop in vertical-lr, spanner");

    test(function() {
        assert_equals(document.getElementById('fifth').offsetLeft, 150);
        assert_equals(document.getElementById('fifth').offsetTop, 0);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, fifth element");

    test(function() {
        assert_equals(document.getElementById('sixth').offsetLeft, 150);
        assert_equals(document.getElementById('sixth').offsetTop, 110);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, sixth element");

    test(function() {
        assert_equals(document.getElementById('seventh').offsetLeft, 150);
        assert_equals(document.getElementById('seventh').offsetTop, 220);
    }, "offsetLeft and offsetTop in vertical-lr at column boundaries, seventh element");
</script>
